requirements.txt
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==3.3.2
colorama==0.4.6
comtypes==1.4.6
filelock==3.15.4
fsspec==2024.6.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.9.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.4.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
openai-whisper==20231117
pillow==10.4.0
PyAudio==0.2.14
pypiwin32==223
pyttsx3==2.91
pywin32==306
regex==2024.7.24
requests==2.32.3
rfc3986==1.5.0
setuptools==74.0.0
sniffio==1.3.1
sympy==1.13.2
tiktoken==0.7.0
torch==2.4.0+cu118
torchaudio==2.4.0+cu118
torchvision==0.19.0+cu118
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
